+++ test/render-pipeline.spec.ts
@@ -36,7 +36,7 @@ function createChart(data: ChartData[]) {
     }
   );
 
-  chart.draw(data);
+  chart.setData(data);
   charts.push(chart);
   return chart;
 }